requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-heroku==0.3.1
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
jwt==1.3.1
MarkupSafe==2.1.2
packaging==23.0
psycopg2==2.9.5
pycparser==2.21
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.4.0